requirements.txt
bugsnag==3.4.2
dj-database-url==0.4.2
Django==2.0.1
django-cors-headers==2.1.0
django-guardian==1.4.9
django-json-widget==0.1.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
awesome-slugify==1.6.5
